refactor: 불필요한 로그 출력 코드 및 주석 삭제
